requirements.txt
PyQt6>=6.6.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
numpy>=1.24.0
